separate phrases and categories into their own classes
